# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Orchestration backend for the automated short-form video pipeline, plus the original book sample endpoints.",
        "title": "Pom Pom Hollow Production API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pipeline/alerts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "What the pipeline has flagged for a person. Defaults to open alerts; pass ?all=true for the history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "List alerts",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include resolved and acknowledged alerts",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Alert"
                            }
                        }
                    }
                }
            }
        },
        "/pipeline/alerts/{id}/ack": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Marks an alert as seen. Distinct from resolved, which the pipeline sets by itself once the condition passes — acknowledging says a person looked, not that the problem went away.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-control"
                ],
                "summary": "Acknowledge an alert",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Alert"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/episodes/{id}/approvals": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-approvals"
                ],
                "summary": "List the gate decisions for an episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ApprovalLog"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "The only way past the two mandatory review gates, and restricted to the reviewer role so an agent cannot approve its own work. GATE_1_SCRIPT releases an approved script into production; GATE_2_RELEASE clears an assembled episode for publishing. The decision is attributed to the authenticated caller, and a rejection sends the episode back for rework.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-approvals"
                ],
                "summary": "Record a human decision at a gate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Gate decision",
                        "name": "approval",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ApprovalLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/reviews/pending": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Powers the reviewer dashboard: every episode currently parked at a gate. Filter to one gate with ?gate=GATE_1_SCRIPT",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-approvals"
                ],
                "summary": "List everything waiting on a human",
                "parameters": [
                    {
                        "enum": [
                            "GATE_1_SCRIPT",
                            "GATE_2_RELEASE"
                        ],
                        "type": "string",
                        "description": "Limit to one gate",
                        "name": "gate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Episode"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Alert": {
            "type": "object",
            "properties": {
                "acknowledged_at": {
                    "type": "string"
                },
                "acknowledged_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "episode_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "stage": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                }
            }
        },
        "main.ApprovalLog": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "decision": {
                    "type": "string"
                },
                "episode_id": {
                    "type": "integer"
                },
                "from_status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "gate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "reviewer": {
                    "type": "string"
                },
                "to_status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                }
            }
        },
        "main.ApprovalRequest": {
            "type": "object",
            "properties": {
                "decision": {
                    "type": "string"
                },
                "gate": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "main.Episode": {
            "type": "object",
            "properties": {
                "ai_disclosure": {
                    "type": "boolean"
                },
                "attempts": {
                    "description": "Attempts counts how many times the current stage has been tried. It\nresets whenever the episode moves on, so it measures this stage rather\nthan the episode's whole history.",
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "string"
                },
                "claimed_by": {
                    "description": "ClaimedBy is the worker currently holding this episode, and ClaimedAt\nis when it took it. The claim is a lease: it expires so a worker that\ndies mid-job does not strand the episode. Both are cleared whenever the\nepisode changes stage.",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "curriculum_topic_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "failed_from": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "id": {
                    "type": "integer"
                },
                "last_error": {
                    "type": "string"
                },
                "made_for_kids": {
                    "description": "MadeForKids drives the YouTube selfDeclaredMadeForKids flag. It defaults\nto true because every episode of this show is children's content, and\ngetting this wrong is a COPPA problem rather than a cosmetic one.",
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "script_text": {
                    "type": "string"
                },
                "song_lyrics": {
                    "type": "string"
                },
                "song_title": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "synopsis": {
                    "type": "string"
                },
                "tiktok_video_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "youtube_video_id": {
                    "type": "string"
                }
            }
        },
...
        "main.EpisodeStatus": {
            "type": "string",
            "enum": [
                "IDEA_BACKLOG",
                "SCRIPT_DRAFT",
                "SCRIPT_APPROVED",
                "VO_GENERATED",
                "MUSIC_GENERATED",
                "ANIMATION_RENDERED",
                "ASSEMBLED",
                "QA_REVIEW",
                "PLATFORM_ADAPTED",
                "SCHEDULED",
                "PUBLISHED",
                "ANALYZED",
                "CANCELLED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusIdeaBacklog",
                "StatusScriptDraft",
                "StatusScriptApproved",
                "StatusVOGenerated",
                "StatusMusicGenerated",
                "StatusAnimationRendered",
                "StatusAssembled",
                "StatusQAReview",
                "StatusPlatformAdapted",
                "StatusScheduled",
                "StatusPublished",
                "StatusAnalyzed",
                "StatusCancelled",
                "StatusFailed"
            ]
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}